Fourth Submit

+++ src/main/java/lang/print/gaps/task5/NumberDivider.java
@@ -3,7 +3,7 @@ package lang.print.gaps.task5;
 public class NumberDivider {
     void divide(int toBeDivided) {
     
-	float result = toBeDivided / 5.0;
+	double result = toBeDivided / 5.0;
 	System.out.println(result);
